requirements.txt
anvil-uplink==0.3.40
anyascii==0.3.0
nltk==3.6.2
PyYAML==5.4.1
stopwords==1.0.0
tqdm==4.62.2
wget==3.2
python-Levenshtein==0.12.2
qwikidata==0.4.0
geopy==2.2.0
